+++ backend/src/main/java/com/project/Tamago/dto/responseDto/LoginResDto.java
@@ -1,0 +1,10 @@
+package com.project.Tamago.dto.responseDto;
+
+import com.project.Tamago.dto.SuccessMessage;
+
+import lombok.Data;
+
+@Data
+public class LoginResDto extends SuccessMessage {
+	private final String accessToken;
+}
