+++ backend/src/main/java/com/project/Tamago/jwt/JwtTokenProvider.java
@@ -25,9 +25,9 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Component;
 
-import com.project.Tamago.dto.TokenDto;
 import com.project.Tamago.exception.CustomException;
 import com.project.Tamago.util.constants.KeyType;
+import com.project.Tamago.util.security.Token;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -66,8 +66,8 @@ public class JwtTokenProvider {
 		expireTimes.put(KeyType.REFRESH, refreshTokenExpireTime);
 	}
 
-	public TokenDto createToken(Authentication authentication) {
-		return new TokenDto(createAccessToken(authentication), createRefreshToken(authentication));
+	public Token createToken(Authentication authentication) {
+		return new Token(createAccessToken(authentication), createRefreshToken(authentication));
 	}
 
 	public String createAccessToken(Authentication authentication) {
